requirements.txt
Django==3.1.2
Pillow==7.2.0
requests==2.24.0